Offload flake input downloads to the recursive derivation (#1)
This PR revamps the recursive build utility provided by this flake. Previously, we were using the `inputs` of the flake object, but the problem was that the dependency tree of a flake is created in such a way that referring to the flakes inside perform the download at Nix evaluation time. That meant that while trying to construct the recursive derivation, we were already downloading them, defeating the whole purpose of reducing the nix evaluation time dependencies.

With this PR, we're pushing the fetching of those dependencies into the recursive derivation using a modified version of `flake-compat` that uses `pkgs.fetchzip` to do the fetch. This is different from building the flake directly with `nix build` or using vanilla `flake-compat` inside the recursive derivation, because then Nix tries to fetch the inputs at nix eval time and fails to perform the download. `pkgs.fetchzip` offloads the downloads to individual derivations, which works fine inside the recursive derivation since then the download gets performed by a regular builder outside the recursive-nix derivation.

* Use a special flake-compat  that fetches as drv

* Bump flake compat to reduce strictness

* Bump flake-compat

* Revamp the recursive build infra

Now the heavy lifting happens inside the recursive derivation by using a
special flake-compat that does the input fetching via derivation.

* Switch to a kadena-io fork of flake-compat